% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_FolderName_to_Files.R
\name{Add_FolderName_to_Files}
\alias{Add_FolderName_to_Files}
\title{Add_FolderName_to_Files}
\usage{
Add_FolderName_to_Files(parent_folder)
}
\arguments{
\item{parent_folder}{Folder with subfolders containing the files to be renamed by adding their folder name.}
}
\value{
All files in subfolders in the parent_folder renamed by adding their folder name.
}
\description{
Adding the folder names to their files
}
\examples{
Add_FolderName_to_Files(parent_folder)
}
