% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_NestedFolderNames_to_Files.R
\name{Add_NestedFolderNames_to_Files}
\alias{Add_NestedFolderNames_to_Files}
\title{Add_NestedFolderNames_to_Files}
\usage{
Add_NestedFolderNames_to_Files(parent_folder)
}
\arguments{
\item{parent_folder}{Folder with nested subfolders containing the files to be renamed by adding their folder names in the path.}
}
\value{
non--All files in subfolders in the parent_folder will be renamed by adding their folder names in the path.
}
\description{
Adding the folder names in the path to their files
}
