% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avasoft_import.R
\name{avasoft_import}
\alias{avasoft_import}
\title{Import and merge csv files from Avasoft software in a single csv file suitable for analysis}
\usage{
avasoft_import(
  folder = getwd(),
  spectrometer = "nir",
  saturation = 5,
  file = "MergedSpectra.csv"
)
}
\arguments{
\item{folder}{a folder with the Avasoft csv files.}

\item{spectrometer}{nir (default) or visible.}

\item{saturation}{the value that should be assigned in case of saturation (default=5).}

\item{file}{the merged file.}
}
\value{
a single csv file with all spectra merged and suitable for analysis.
}
\description{
Import and merge csv files from Avasoft software in a single csv file suitable for analysis
}
