% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copy_From_SubFolders.R
\name{copy_from_subfolders}
\alias{copy_from_subfolders}
\title{Add_FolderName_to_Files}
\usage{
copy_from_subfolders(parent_folder, new_folder, recurse = FALSE)
}
\arguments{
\item{parent_folder}{Folder with subfolders containing the files to be copied.}

\item{new_folder}{the Folder to contain the copied files.}

\item{recurse}{to specify if it recurse or not (default recurse=FALSE)}
}
\value{
All the files to be copied in the new_folder
}
\description{
Adding the folder names to their files
}
\examples{
copy_from_subfolders(parent_folder, new_folder, recurse=FALSE)
}
