% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emsch.R
\name{emsch}
\alias{emsch}
\title{emsch}
\usage{
emsch(spectra)
}
\arguments{
\item{spectra}{a hyperspec object.}
}
\value{
a hyperspec object of emsc preprocessed spectra.
}
\description{
preprocess spectra in hyperspec formate by emsc
}
\examples{
emsch(spectra)
}
