% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_and_Merge_Spectra.R
\name{Import_and_Merge_Spectra}
\alias{Import_and_Merge_Spectra}
\title{Import_and_Merge_Spectra}
\usage{
Import_and_Merge_Spectra(
  folder = getwd(),
  file = "MergedSpectra.csv",
  delimiter = ",",
  file_extension = ".CSV",
  r_Skip = 0
)
}
\arguments{
\item{folder}{a folder with the txt/csv files with spectral data.}

\item{file}{the merged file.}

\item{delimiter}{The delimiter used in the files with spectral data.}

\item{file_extension}{The extension of the spectral data files. Default is ".CSV"}

\item{r_Skip}{number of lines to be skipped before start reading (skipping initial metadata for instance). Default is 0}
}
\value{
a single csv file with all spectra merged and suitable for analysis.
}
\description{
Import and merge csv files of spectral data in a single csv file suitable for analysis. Suitable for files with only 2 columns: 1 for the Wavelength/ Wavenumber and the other with the absorbance,transmittance or reflectance
}
