% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_subfiles.R
\name{Merge_subfiles}
\alias{Merge_subfiles}
\title{Merge_subfiles}
\usage{
Merge_subfiles(parent_folder, new_folder, recurse = FALSE)
}
\arguments{
\item{parent_folder}{Folder with subfolders containing the subfiles to be merged.}

\item{new_folder}{Folder for the subfiles to be copied to.}
}
\value{
All files copied from subfolders in the parent_folder to the new_folder.
}
\description{
Copying files from sub-folders in one folder
}
\examples{
Merge_subfiles(parent_folder, new_folder, recurse=FALSE)
}
