% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegCV.R
\name{RegCV}
\alias{RegCV}
\title{RegCV}
\usage{
RegCV(M, V, comp = 20, model = "pls", validation = "repeatedcv")
}
\arguments{
\item{M}{a hyperspec object}

\item{V}{the traget variable}

\item{comp}{maximum number of principle components}

\item{model}{the type of the model (default=pls).}

\item{validation}{the type of the CV (default=repeatedcv).}
}
\value{
The R2 and and best tune parameters
}
\description{
Perform crossvalidation for regression models (This version need further revision)
}
