% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvh.R
\name{snvh}
\alias{snvh}
\title{snvh}
\usage{
snvh(spectra)
}
\arguments{
\item{spectra}{a hyperspec object.}
}
\value{
a hyperspec object of SNV preprocessed spectra.
}
\description{
Preprocess spectra in a hyperspec format by SNV
}
\examples{
snvh(spectra)
}
