% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecImp.R
\name{SpecImp}
\alias{SpecImp}
\title{SpecImp}
\usage{
SpecImp(x, y = 8)
}
\arguments{
\item{x}{a csv file.}

\item{y}{the number of lines to be skipped.}
}
\value{
a dataframe of the imported spectra.
}
\description{
read csv files separated with ";" in R
}
\examples{
ImportedFiles<- lapply (files, SpecImp)
}
