% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraPreProcess.R
\name{SpectraPreProcess}
\alias{SpectraPreProcess}
\title{SpectraPreProcess}
\usage{
SpectraPreProcess(
  rawspectra,
  SmLn = 3,
  SmMn = 5,
  SmHn = 11,
  ExtraDerivativeLSm = 15,
  ExtraDerivativeHSm = 17
)
}
\arguments{
\item{rawspectra}{a hyperspec object of the spectra to be prepossessed.}

\item{SmLn}{Numeric. The filter length of the low level of smoothing to be attempted.}

\item{SmHn}{Numeric. The filter length of the high level of smoothing to be attempted.}

\item{ExtraDerivativeLSm}{Numeric. The filter length of the low level of smoothing to be attempted for both the 1st and 2nd derivatives in addition to the default values (3).}

\item{ExtraDerivativeHSm}{Numeric. The filter length of the high level of smoothing to be attempted for both the 1st and 2nd derivatives in addition to the default values (9,11).}
}
\value{
A list of 66 preprocessed spectra and a list of their names.
}
\description{
create a list of all spectra after 66 diverse preprocessing
}
\examples{
SpectraPreProcess(spectra,SmLn=3,SmMn=5,SmHn=11, ExtraDerivativeLSm=15, ExtraDerivativeHSm=17)
}
