% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraPreProcessE.R
\name{SpectraPreProcessE}
\alias{SpectraPreProcessE}
\title{SpectraPreProcessE}
\usage{
SpectraPreProcessE(rawspectra, SmLn = 3, SmMn = 5, SmHn = 11)
}
\arguments{
\item{rawspectra}{a hyperspec object of the spectra to be prepossessed.}

\item{SmLn}{Numeric. The filter length of the low level of smoothing to be attempted.}

\item{SmHn}{Numeric. The filter length of the high level of smoothing to be attempted.}
}
\value{
A list of 74 preprocessed spectra and a list of their names these include preprocessing by emsc, which can cause data leakage in case of using the set in Crossvalidation; in such case use SpectraPreProcess function.
}
\description{
create a list of all spectra after 74 diverse preprocessing
}
\examples{
SpectraPreProcessE(spectra,SmLn=3,SmMn=5,SmHn=11)
}
