% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraSmoothing.R
\name{SpectraSmoothing}
\alias{SpectraSmoothing}
\title{SpectraSmoothing}
\usage{
SpectraSmoothing(rawspectra, filter_lengths = seq(3, 20, by = 2))
}
\arguments{
\item{rawspectra}{a hyperspec object.}

\item{filter_lengths}{a list of filter lengths to be attempted for smoothing}
}
\value{
A list of preprocessed spectra.
}
\description{
create a list of all spectra after 74 diverse preprocessing
}
