% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnName.R
\name{tnName}
\alias{tnName}
\title{tnName}
\usage{
tnName(fx)
}
\arguments{
\item{fx}{a dataframe}
}
\value{
transposed dataframe with 1st row of the original dataframe as the new column names
}
\description{
transpose a df and make the 1st row the new column names
}
