% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvnormalize.R
\name{uvnormalize}
\alias{uvnormalize}
\title{uvnormalize}
\usage{
uvnormalize(x)
}
\arguments{
\item{x}{a hyperspec object.}
}
\value{
a hyperspec object of unit vector normalized preprocessed spectra.
}
\description{
unit vector normalization
}
\examples{
uvnormalize(x)
}
