% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSCV.R
\name{PLSCV}
\alias{PLSCV}
\title{PLSCV}
\usage{
PLSCV(M, V, ncompMax = 20, validation = "repeatedcv")
}
\arguments{
\item{M}{a hyperspec object}

\item{V}{the traget variable}

\item{validation}{the type of the CV (default=repeatedcv).}
}
\value{
The R2 and and best tune parameters
}
\description{
Perform crossvalidation for PLS regression (This version need further revision)
}
\examples{
PLSCV(M,V,ncompMax=20,validation="repeatedcv")
}
