\name{uvnormalize}
\alias{uvnormalize}
\title{uvnormalize}
\usage{
uvnormalize(spectra)
}
\description{
preprocess spectra in hyperspec formate by emsc
}
\arguments{
\item{spectra}{A Hyperspec object.}
}
\value{
a hyperspec object of unit vector normalized preprocessed spectra.
}
\examples{
uvnormalize(spectra)
}
